Revert "Fix fr truncated boot image size recorded with ElTorito boot catalog and boot info table."
This reverts commit fc7fcfd14cdf376e7cfe108d827d3e9a0d61d8fd.

+++ filesystem/iso9660/finalize.go
@@ -594,7 +594,7 @@ func (fsm *FileSystem) Finalize(options FinalizeOptions) error {
 				return fmt.Errorf("unable to find image child %s: %v", e.BootFile, err)
 			}
 			// save the child so we can add location late
-			e.size = uint32(child.size)
+			e.size = uint16(child.size)
 			child.elToritoEntry = e
 		}
 	}
